rearange infer_util
