+++ numpyro/infer_util.py
@@ -1,3 +1,30 @@
+import jax.numpy as np
+
+from numpyro.handlers import substitute, trace
+
+
+def log_density(model, model_args, model_kwargs, params):
+    """
+    Computes log of joint density for the model given latent values ``params``.
+
+    :param model: Python callable containing Pyro primitives.
+    :param tuple model_args: args provided to the model.
+    :param dict model_kwargs`: kwargs provided to the model.
+    :param dict params: dictionary of current parameter values keyed by site
+        name.
+    :return: log of joint density and a corresponding model trace
+    """
+    model = substitute(model, params)
+    model_trace = trace(model).get_trace(*model_args, **model_kwargs)
+    log_joint = 0.
+    for site in model_trace.values():
+        if site['type'] == 'sample':
+            log_prob = np.sum(site['fn'].log_prob(site['value']))
+            if 'scale' in site:
+                log_prob = site['scale'] * log_prob
+            log_joint = log_joint + log_prob
+    return log_joint, model_trace
+
 
 def transform_fn(transforms, params, invert=False):
     """
